+++ docsource/releasenotes.rst
@@ -1,8 +1,12 @@
 Workspace service release notes
 ===============================
 
-VERSION: 0.4.2 (Released TBD)
+VERSION: 0.5.0 (Released TBD)
 -----------------------------
+
+BACKWARDS INCOMPATIBILITIES:
+
+* The ``skip`` parameter of ``list_objects`` has been removed.
 
 UPDATED FEATURES / MAJOR BUG FIXES: